Moved all types into their new homes and fixed all references. Stuff is still tightly coupled, but we'll fix that along the way!
