--- SharpCrokite.Core/App.xaml
@@ -1,8 +1,0 @@
-﻿<Application x:Class="MyEveToolset.App"
-             xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-             xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             Startup="Startup_App">
-    <Application.Resources>
-         
-    </Application.Resources>
-</Application>
