+++ SharpCrokite.Core/StaticDataUpdater/JSONModels/GroupJSON.cs
@@ -1,4 +1,4 @@
-﻿namespace MyEveToolset.StaticDataUpdater.JSONModels
+﻿namespace SharpCrokite.Core.StaticDataUpdater.JSONModels
 {
     public class GroupJSON : IGroupJSON
     {
