+++ SharpCrokite.Core/StaticDataUpdater/JSONModels/ICategoryJSON.cs
@@ -1,4 +1,4 @@
-﻿namespace MyEveToolset.StaticDataUpdater.JSONModels
+﻿namespace SharpCrokite.Core.StaticDataUpdater.JSONModels
 {
     public interface ICategoryJSON
     {
