--- SharpCrokite.DataAccess/Class1.cs
@@ -1,8 +1,0 @@
-﻿using System;
-
-namespace SharpCrokite.DataAccess
-{
-    public class Class1
-    {
-    }
-}
