--- SharpCrokite.Infrastructure/Class1.cs
@@ -1,8 +1,0 @@
-﻿using System;
-
-namespace SharpCrokite.Infrastructure
-{
-    public class Class1
-    {
-    }
-}
