+++ SharpCrokite.UI/App.xaml
@@ -1,8 +1,7 @@
 ﻿<Application x:Class="SharpCrokite.UI.App"
              xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
              xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-             xmlns:local="clr-namespace:SharpCrokite.UI"
-             StartupUri="MainWindow.xaml">
+             Startup="Startup_App">
     <Application.Resources>
          
     </Application.Resources>
