+++ SharpCrokite.UI/MainWindow.xaml
@@ -1,12 +1,77 @@
-﻿<Window x:Class="SharpCrokite.UI.MainWindow"
+﻿<Window x:Class="SharpCrokite.UI.SharpCrokiteMainWindow"
         xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
         xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
         xmlns:d="http://schemas.microsoft.com/expression/blend/2008"
         xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
-        xmlns:local="clr-namespace:SharpCrokite.UI"
-        mc:Ignorable="d"
-        Title="MainWindow" Height="450" Width="800">
-    <Grid>
-
+        mc:Ignorable="d"  ResizeMode="NoResize"
+        Title="Sharp Crokite" Height="1080" Width="1920"
+        WindowStartupLocation="CenterScreen" Icon="Images/sharp-crokite-icon.png">
+    <Grid x:Name="MainGrid">
+        <TabControl>
+            <TabItem Header="Mining Yields">
+                <DataGrid Margin="10,10,10,10"></DataGrid>
+            </TabItem>
+            <TabItem Header="Market Data">
+                <Grid>
+                    <StackPanel Orientation="Horizontal" Height="45" VerticalAlignment="Top">
+                        <Button Content="Update Static Data" Name="UpdateStaticDataButton" Click="UpdateStaticDataButton_Click" Width="160" Margin="10,10,0,5"/>
+                        <Button Content="Delete Static Data" Name="DeleteStaticDataButton" Click="DeleteStaticDataButton_Click" Width="160" Margin="5,10,0,5"/>
+                        <Button Content="Update Prices" Name="UpdatePriceDataButton" Click="UpdatePriceDataButton_Click" Width="160" Margin="5,10,0,5"/>
+                        <Button Content="Delete Prices" Name="DeletePriceDataButton" Click="DeletePriceDataButton_Click" Width="160" Margin="5,10,0,5"/>
+                    </StackPanel>
+                    <DataGrid Name="HarvestablesDataGrid" Margin="10,50,10,10" IsReadOnly="true" AutoGenerateColumns="False" 
+                              HorizontalGridLinesBrush="LightGray" VerticalGridLinesBrush="LightGray" HeadersVisibility="Column"  
+                              CanUserAddRows="False" CanUserDeleteRows="False" CanUserResizeRows="False" CanUserSortColumns="False"
+                              CanUserReorderColumns="False" CanUserResizeColumns="False">
+                        <DataGrid.CellStyle>
+                            <Style TargetType="{x:Type DataGridCell}">
+                                <Style.Triggers>
+                                    <Trigger Property="DataGridCell.IsSelected" Value="True">
+                                        <Setter Property="BorderBrush" Value="Transparent"/>
+                                        <Setter Property="Foreground" Value="{DynamicResource {x:Static SystemColors.ControlTextBrushKey}}"/>
+                                        <Setter Property="Background" Value="Transparent"/>
+                                    </Trigger>
+                                </Style.Triggers>
+                            </Style>
+                        </DataGrid.CellStyle>
+                        <DataGrid.Columns>
+                            <DataGridTemplateColumn  Header="Icon">
+                                <DataGridTemplateColumn.CellTemplate>
+                                    <DataTemplate>
+                                        <Image Source="{Binding Icon}" Width="48" Height="48" ToolTip="{Binding Description}" ToolTipService.InitialShowDelay="100" />
+                                    </DataTemplate>
+                                </DataGridTemplateColumn.CellTemplate>
+                            </DataGridTemplateColumn>
+                            <DataGridTextColumn Header="Name" Binding="{Binding Name}" Width="240">
+                                <DataGridTextColumn.ElementStyle>
+                                    <Style TargetType="TextBlock">
+                                        <Setter Property="VerticalAlignment" Value="Center" />
+                                        <Setter Property="Margin" Value="5,5,5,5" />
+                                    </Style>
+                                </DataGridTextColumn.ElementStyle>
+                            </DataGridTextColumn>
+                            <DataGridTextColumn Header="Price" Binding="{Binding Price}"  Width="160">
+                                <DataGridTextColumn.ElementStyle>
+                                    <Style TargetType="TextBlock">
+                                        <Setter Property="VerticalAlignment" Value="Center" />
+                                        <Setter Property="Margin" Value="5,5,5,5" />
+                                    </Style>
+                                </DataGridTextColumn.ElementStyle>
+                            </DataGridTextColumn>
+                            <DataGridTextColumn Header="Material Content" Binding="{Binding MaterialContents}" Width="240">
+                                <DataGridTextColumn.ElementStyle>
+                                    <Style TargetType="TextBlock">
+                                        <Setter Property="VerticalAlignment" Value="Center" />
+                                        <Setter Property="Margin" Value="5,5,5,5" />
+                                    </Style>
+                                </DataGridTextColumn.ElementStyle>
+                            </DataGridTextColumn>
+                        </DataGrid.Columns>
+                    </DataGrid>
+                </Grid>
+            </TabItem>
+            <TabItem Header="Settings">
+            </TabItem>
+        </TabControl>
     </Grid>
 </Window>
